code/requirements.txt
attrs==22.1.0
black==22.8.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Deprecated==1.2.13
idna==3.4
jsonschema==4.16.0
mypy-extensions==0.4.3
pathspec==0.10.1
pdfminer==20191125
platformdirs==2.5.2
publicsuffixlist==0.8.0
pycryptodome==3.15.0
pymisp==2.4.162
pyrsistent==0.18.1
python-dateutil==2.8.2
PyYAML==6.0
requests==2.28.1
six==1.16.0
tomli==2.0.1
urllib3==1.26.12
wrapt==1.14.1
